extract where clause types
